create migration, seeder, factory, model for tbl Users
